refactor(core): split projection matrix in renderer
